requirements.txt
pywin32
pyyaml